requirements.txt
PySide6==6.6.0
Pillow==10.0.1
mss==9.0.1
pynput==1.7.6
easyocr==1.7.0
numpy==1.24.3
requests==2.31.0
opencv-python==4.8.1.78